standalone-processing/requirements.txt
numpy==1.22.3
tqdm
matplotlib
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.12.0
torchvision==0.13.0
hydra-core==1.0.5
scikit-image
imgviz==1.5.0
segmentation-models-pytorch==0.2.1
pytorch-lightning==1.6.5
geopy==2.2.0
pandas==1.4.2
exif==1.3.5
opencv-python==4.5.4.60
openpyxl==3.0.9
piexif==1.1.3
Pillow==9.0.1
